Relax transcript-to-gene joining and extract TranscriptToGeneJoiner
